flask_app/requirements.txt
flask==3.0.2
numpy==1.26.4
pandas==2.2.1
Pillow==10.2.0
open-clip-torch==2.29.0
torch==2.2.1
torchvision==0.17.1
scikit-learn==1.4.1
